requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==5.1.1
django-redis==5.4.0
djangorestframework==3.15.2
jwt==1.3.1
psycopg2-binary==2.9.9
pycparser==2.22
redis==5.0.8
setuptools==75.1.0
sqlparse==0.5.1
wheel==0.44.0
